% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsadd.R
\name{tsadd}
\alias{tsadd}
\title{Adds two TS objects together}
\usage{
tsadd(series1, series2)
}
\arguments{
\item{series1}{A ts object}

\item{series2}{A ts object}
}
\value{
The new ts object.
}
\description{
Adds two TS objects together
}
\examples{
temp <- c(1:14)
Series1 <- ts(temp,start=c(2016,1),frequency = 12)
Series2 <- ts(temp,start=c(2016,1),frequency = 12)
SeriesA <- tsadd(series1=Series1,series2=Series2)
}
