% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslength.R
\name{tslength}
\alias{tslength}
\title{Returns the length of a ts object}
\usage{
tslength(series)
}
\arguments{
\item{series}{A ts object}
}
\value{
The number of periods in the ts object.
}
\description{
Returns the length of a ts object
}
